add: test global limit (default)
